az204pythonfunctionappdemo/pyproject.toml
[tool.poetry]
name = "az204_python_functionapp_demo"
version = "0.1.0"
description = ""
packages = [{ include = "HttpExample" }]
authors = []

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.0"
loguru = "^0.7.2"
requests = "^2.32.3"
azure-functions = "^1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


